requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
langchain>=0.2.5
langchain-openai>=0.1.0
langchain-community>=0.2.5
llama-index==0.10.43
llama-index-llms-openai==0.1.22
llama-index-embeddings-openai==0.1.10
llama-index-vector-stores-chroma==0.1.10
openai==1.30.5
chromadb==0.5.3
slowapi==0.1.9
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
aiosqlite==0.20.0
python-dotenv==1.0.1
numpy==1.26.4
